refactoring the admin code
