+++ src/config/config.json
@@ -1,9 +1,9 @@
 {
   "development": {
-    "username": "branchxdb_owner",
-    "password": "npg_jTXwO8a3AufG",
-    "database": "branchxdb",
-    "host": "ep-white-breeze-abravfmm-pooler.eu-west-2.aws.neon.tech",
+    "username": "neondb_owner",
+    "password": "npg_XMKY7Sgnkd0P",
+    "database": "branchx",
+    "host": "ep-flat-boat-abu0yaw0-pooler.eu-west-2.aws.neon.tech",
     "dialect": "postgres",
     "dialectOptions": {
       "ssl": {
@@ -20,10 +20,10 @@
     "dialect": "mysql"
   },
   "production": {
-    "username": "branchxdb_owner",
-    "password": "npg_jTXwO8a3AufG",
-    "database": "branchxdb",
-    "host": "ep-white-breeze-abravfmm-pooler.eu-west-2.aws.neon.tech",
+    "username": "neondb_owner",
+    "password": "npg_XMKY7Sgnkd0P",
+    "database": "branchx",
+    "host": "ep-flat-boat-abu0yaw0-pooler.eu-west-2.aws.neon.tech",
     "dialect": "postgres",
      "dialectOptions": {
       "ssl": {
@@ -31,6 +31,5 @@
         "rejectUnauthorized": false
       }
     }
-
   }
 }
