+++ src/middlewares/auth-middleware.js
@@ -2,7 +2,7 @@ const jwt = require('jsonwebtoken');
 const { ServerConfig } = require('../config');
 
 const isOrgAdmin = (req, res, next) => {
-  if (req.user.role !== 'ORG_ADMIN') {
+  if (req.user.role !== 'SUPERADMIN') {
     return res.status(403).json({ message: 'Only Org Admins are allowed' });
   }
   next();
@@ -10,7 +10,7 @@ const isOrgAdmin = (req, res, next) => {
 
 const isAdmin = (req, res, next) => {
 
-  if (!req.user || !['ADMIN', 'ORG_ADMIN'].includes(req.user.role)) {
+  if (!req.user || !['ADMIN', 'SUPERADMIN'].includes(req.user.role)) {
     return res.status(403).json({ message: 'Access denied. Not an admin.' });
   }
   next();
