+++ src/repositories/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+    AdminRepository :require('./admin-repository'),
+    CampaignRepository : require('./campaign-repository')
+}
